Add 404, 500, util
