+++ packages/nextjs-template/pages/404.tsx
@@ -1,0 +1,7 @@
+import Error from "next/error";
+
+export { getStaticProps } from "./util";
+
+export default function Custom404() {
+  return <Error statusCode={404} />;
+}
