+++ packages/nextjs-template/pages/500.tsx
@@ -1,0 +1,7 @@
+import Error from "next/error";
+
+export { getStaticProps } from "./util";
+
+export default function Custom500() {
+  return <Error statusCode={500} />;
+}
